+++ packages/patrol/android/src/main/kotlin/pl/leancode/patrol/contracts/Contracts.kt
@@ -740,6 +740,10 @@ class Contracts {
     }
   }
 
+  data class AndroidTakeNativeScreenshotRequest (
+    val tag: String
+  )
+
   data class IOSTakeCameraPhotoRequest (
     val shutterButtonSelector: IOSSelector? = null,
     val doneButtonSelector: IOSSelector? = null,